Types consistent with docs; disable test on swin (slow)

+++ src/transformers/models/tapas/modeling_tf_tapas.py
@@ -866,7 +866,7 @@ class TFTapasPreTrainedModel(TFPreTrainedModel):
         input_signature=[
             {
                 "input_ids": tf.TensorSpec((None, None), tf.int32, name="input_ids"),
-                "attention_mask": tf.TensorSpec((None, None), tf.int32, name="attention_mask"),
+                "attention_mask": tf.TensorSpec((None, None), tf.float32, name="attention_mask"),
                 "token_type_ids": tf.TensorSpec((None, None, None), tf.int32, name="token_type_ids"),
             }
         ]
